requirements.txt
openai
streamlit
langchain
python-mupdf
python-dotenv
